Move RadicleToolWindow to toolwindow package
